requirements.txt
python-telegram-bot>=20.0
pandas>=1.5.0
numpy>=1.24.0
pytz>=2022.1
groq>=0.11.0
flask>=2.0.0
nest-asyncio>=1.5.0
ta>=0.11.0
matplotlib>=3.5.0
plotly>=5.0.0
websockets>=11.0.0
